コード整理

+++ behaviorDodge.h
@@ -5,7 +5,7 @@ class BehaviorDodge : public BehaviorNode
 {
 private:
 	bool m_isDodge = false;
-	int m_DodgeFrame = 0.0f;
+	float m_DodgeFrame = 0.0f;
 
 public:
 	using BehaviorNode::BehaviorNode;
